app/requirements.txt
konlpy
mecab-python
fastapi
uvicorn[standard]
selenium
google-cloud-translate==2.0.1